🚚 Reorganize components, pages, and utils * Split utils.ts up into separate utils within a /utils directory * Rename Home page file to Home.tsx * Remove CARGO logo * Create /assets directory, add new logo * Update import paths to reflect new filenames, take advantage of new config settings
