+++ src/components/NewRunFormModal.tsx
@@ -18,7 +18,7 @@
 
 import React from "react";
 import Button from "@icgc-argo/uikit/Button";
-import { ModalPortal } from "../App";
+import { ModalPortal } from "App";
 import Modal from "@icgc-argo/uikit/Modal";
 import InputLabel from "@icgc-argo/uikit/form/InputLabel";
 import Input from "@icgc-argo/uikit/form/Input";
@@ -28,7 +28,7 @@ import "ace-builds/src-noconflict/theme-github";
 import { css } from "emotion";
 import { ApolloError } from "apollo-client";
 import { useTheme } from "@icgc-argo/uikit/ThemeProvider";
-import { runWorkflow } from "../rdpc";
+import { runWorkflow } from "rdpc";
 
 type RunSuccess = {
   run_id: string;
