+++ src/components/RDPCStats.tsx
@@ -19,7 +19,7 @@
 import React from "react";
 import { css } from "emotion";
 import { Link } from "react-router-dom";
-import { DashboardTask, RunCompact } from "../gql/types";
+import { DashboardTask, RunCompact } from "gql/types";
 import Typography from "@icgc-argo/uikit/Typography";
 import { useTheme } from "@icgc-argo/uikit/ThemeProvider";
