+++ src/components/RunsTable.tsx
@@ -19,13 +19,13 @@
 import React from "react";
 import Table, { TableColumnConfig } from "@icgc-argo/uikit/Table";
 import { Link } from "react-router-dom";
-import { DashboardQueryResponse, RunCompact } from "../gql/types";
-import { parseEpochToEST } from "../utils";
-import { CancelConfirmModal, CancelResponseModal, CancelResponse } from "./CancelRun";
+import { DashboardQueryResponse, RunCompact } from "gql/types";
+import { parseEpochToEST } from "utils/time";
+import { CancelConfirmModal, CancelResponseModal, CancelResponse } from "components/CancelRun";
 import Modal from "@icgc-argo/uikit/Modal";
 import { ApolloError } from "apollo-boost";
 import Button from "@icgc-argo/uikit/Button";
-import { cancelWorkflow } from "../rdpc";
+import { cancelWorkflow } from "rdpc";
 
 export default ({
   runs,
