+++ src/utils/time.ts
@@ -17,7 +17,6 @@
  */
 
 import moment from "moment-timezone";
-import { Task } from "./gql/types";
 
 moment.updateLocale("en", {
   invalidDate: "N/A",
@@ -27,22 +26,3 @@ export const parseEpochToEST = (milli: string) =>
   moment(parseInt(milli))
     .tz("America/Toronto")
     .format("MMMM Do YYYY, h:mm:ss a");
-
-export const sortTasks = (tasks: Task[], reverse = false) => {
-  const stateOrder = [
-    "UNKNOWN",
-    "QUEUED",
-    "RUNNING",
-    "COMPLETE",
-    "EXECUTOR_ERROR",
-  ];
-
-  const sortedTasks = tasks.sort((a, b) => {
-    const ai = stateOrder.indexOf(a.state);
-    const bi = stateOrder.indexOf(b.state);
-
-    return ai > bi ? -1 : ai < bi ? 1 : 0;
-  });
-
-  return reverse ? sortedTasks.reverse() : sortedTasks;
-};
